Add authç

+++ src/controllers/task.controllers.ts
@@ -7,15 +7,17 @@ import Task from '../models/Task'
 import { taskCreateSchema, taskUpdateSchema } from '../schemas/task.schemas'
 
 async function getAllTasks (ctx: Context) {
-  const tasks = await Task.find()
+  const userId = ctx.state.user.id
+  const tasks = await Task.find({ userId })
   const jsonTasks = tasks.map(task => task.toJSON())
   ctx.body = jsonTasks
 }
 
 async function getTask (ctx: Context) {
   const id = ctx.params.id as string
+  const userId = ctx.state.user.id
 
-  const task = await Task.findOne({ id })
+  const task = await Task.findOne({ id, userId })
 
   if (task) {
     ctx.body = task.toJSON()
@@ -26,13 +28,15 @@ async function getTask (ctx: Context) {
 
 async function createTask (ctx: Context) {
   const payload = ctx.request.body as TaskInputDTO
+  const userId = ctx.state.user.id
 
   try {
     const validated: TaskInputDTO = await taskCreateSchema.validateAsync(payload)
 
     const newTask = new Task({
       id: nanoid(),
-      ...validated
+      ...validated,
+      userId
     })
     const response = await newTask.save()
     ctx.body = response
@@ -49,6 +53,7 @@ async function createTask (ctx: Context) {
 async function updateTask (ctx: Context) {
   const id = ctx.params.id as string
   const payload = ctx.request.body as Partial<TaskInputDTO>
+  const userId = ctx.state.user.id
 
   if (Object.values(payload).length === 0) {
     throw new ServiceError(400, 'No data to update')
@@ -57,7 +62,7 @@ async function updateTask (ctx: Context) {
   try {
     const validated = await taskUpdateSchema.validateAsync(payload)
 
-    const updatedTask = await Task.findOneAndUpdate({ id }, validated, { new: true })
+    const updatedTask = await Task.findOneAndUpdate({ id, userId }, validated, { new: true })
 
     if (updatedTask) {
       ctx.body = updatedTask.toJSON()
@@ -74,8 +79,9 @@ async function updateTask (ctx: Context) {
 }
 
 async function deleteTask (ctx: Context) {
+  const userId = ctx.state.user.id
   const id = ctx.params.id as string
-  await Task.findOneAndDelete({ id })
+  await Task.findOneAndDelete({ id, userId })
   ctx.status = 204
 }
